requirements.txt
django
lxml
django-ninja
